add basic ui state to health care screens

+++ app/src/main/java/com/damanhour/Graduation/medisupport/ui/uiElement/navigation/child/activity/ActivityNavGraph.kt
@@ -44,7 +44,7 @@ internal fun NavHostController.popActivityNavGraph() {
 internal fun NavGraphBuilder.activityNavGraph(
     popActivityNavGraph: () -> Unit,
     popHistoryDestination: () -> Unit,
-    navigateToHistoryDestination: () -> Unit,
+    navigateToHistoryDestination: (Int) -> Unit,
     navigateToHeartRateNavGraph: () -> Unit,
     navigateToBmiNavGraph: () -> Unit,
     navigateToBloodSugarNavGraph: () -> Unit,
