+++ app/src/main/java/com/damanhour/Graduation/medisupport/ui/uiElement/screens/activity/ActivityDestination.kt
@@ -15,7 +15,7 @@ const val ACTIVITY_DESTINATION_ROUTE = "activityDestination"
 @RequiresApi(Build.VERSION_CODES.O)
 fun NavGraphBuilder.activityDestination(
     popActivityNavGraph: () -> Unit,
-    navigateToHistoryDestination: () -> Unit,
+    navigateToHistoryDestination: (Int) -> Unit,
     navigateToHeartRateNavGraph: () -> Unit,
     navigateToBloodPressureNavGraph: () -> Unit,
     navigateToBmiNavGraph: () -> Unit,
